requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
gunicorn==22.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
nltk==3.8.1
numpy==1.26.4
packaging==24.0
pyspellchecker==0.7.2
regex==2024.5.10
scikit-learn==1.4.2
scipy==1.13.0
textblob==0.18.0.post0
threadpoolctl==3.5.0
tqdm==4.66.4
Werkzeug==3.0.3
